% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accounting.R
\name{accounting}
\alias{accounting}
\alias{accounting_read}
\alias{accounting_parse}
\title{Extract and parse the SGE job accounting information}
\usage{
accounting(job_ids, tz = "EST")

accounting_read(job_ids)

accounting_parse(accounting_info, tz = "EST")
}
\arguments{
\item{job_ids}{A \code{character} vector of SGE job IDs to inspect.}

\item{tz}{Time zone for your SGE cluster. Defaults to \code{EST} for JHPCE.}

\item{accounting_info}{A named list where each element contains the output
from \code{qacct} for a given job and the names of the list are the SGE job IDs.
This can be produced using \code{\link[=accounting_read]{accounting_read()}}.}
}
\value{
\code{accounting_read}: A list with the output from \code{qacct} by SGE for
each of the SGE job IDs.

\code{accounting_parse} and \code{accounting}: a \code{data.frame} with the SGE
accounting information parsed.
}
\description{
For a given set of SGE job IDs, these functions extract the data from SGE
using \code{qacct} and then parse it to produce a data.frame with the summary
information recorded by SGE on the performance of these SGE jobs.
}
\examples{

## Requires JHPCE to run
if(!is.travis()) accounting(c('92500', '77672'))



## Requires JHPCE access
if(!is_travis()) acc_info_jhpce <- accounting_read('92500')

## The above is identical
acc_info <- list('92500' = readLines(
    system.file('extdata', 'accounting', '92500.txt',
    package = 'sgejobs')))
acc_info

if(!is.travis()) identical(acc_info_jhpce, acc_info)


## Example for an array job
if(!is_travis()) acc_info_jhpce_array <- accounting_read('77672')

## The example file has been subset to just the first two tasks
acc_info_array <- list('77672' = readLines(
    system.file('extdata', 'accounting', '77672.txt',
    package = 'sgejobs')))


## Requires JHPCE access
if(!is_travis()) accounting_info_jhpce <- accounting_read(c('92500', '77672'))

## Here we use the data included in the package to avoid depending on JHPCE
## where the data for job 77672 has been subset for the first two tasks.
accounting_info <- list(
    '92500' = readLines(system.file('extdata', 'accounting', '92500.txt',
        package = 'sgejobs')),
    '77672' = readLines(system.file('extdata', 'accounting', '77672.txt',
        package = 'sgejobs'))
)

## Here we parse the data from `qacct` into a data.frame
res <- accounting_parse(accounting_info)

## Check the maximum memory use
res$maxvmem

## And the absolute maximum
pryr:::show_bytes(max(res$maxvmem))

}
