% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_loop.R
\name{job_loop}
\alias{job_loop}
\title{Build a bash script that loops over variables and submits SGE jobs}
\usage{
job_loop(loops, name, create_shell = FALSE, queue = "shared",
  memory = "10G", cores = 1L, email = "e", logdir = "logs",
  filesize = "100G", task_num = NULL, tc = 20)
}
\arguments{
\item{loops}{A named \code{list} where each of the elements are character vectors.
The names of \code{loops} specify the variables used for the loops and the
contents specify the options to loop through for each variable.}

\item{name}{A \code{character(1)} vector with the name of the script. Any spaces
will be replaced by underscores.}

\item{create_shell}{A \code{logical(1)} vector specifying whether to create a
shell file for the script.}

\item{queue}{A \code{character(1)} vector with the name of the SGE queue. Check
how busy a given queue is by running \code{qpic -q queuename}.}

\item{memory}{The amount of memory per core to request in SGE syntax. You
can check how much a current job is utilizing using the \code{qmem} JHPCE command.
For more detail on the memory options, check
\url{https://jhpce.jhu.edu/knowledge-base/how-to/#MemSpec}.}

\item{cores}{The number of cores to request. Note that the total memory
your job will request is \code{cores} multiplied by \code{memory}.}

\item{email}{The email reporting option for the email. For more information
check \url{https://jhpce.jhu.edu/knowledge-base/how-to/#Email}.}

\item{logdir}{The directory for the SGE log files relative to the current
working directory.}

\item{filesize}{The maximum file size in SGE format.}

\item{task_num}{The number of tasks for your job, which will make it into an
array job. If \code{NULL} this is ignored.}

\item{tc}{If \code{task_num} is specified, this option controls the number of
concurrent tasks.}
}
\value{
A character vector with the script contents. If \code{create_shell} was
specified then it also creates the actual script in the current
working directory.
}
\description{
This function builds a bash script that loops over a set of variables
with pre-specified values to create an internal bash script that then
gets submitted as a SGE job.
}
\examples{

job_loop(
    loops = list(region = c('DLPFC', 'HIPPO'), feature = c('gene', 'exon', 'tx', 'jx')),
    name = 'bsp2_test'
)

job_loop(
    loops = list(region = c('DLPFC', 'HIPPO'), feature = c('gene', 'exon', 'tx', 'jx')),
    cores = 5,
    task_num = 10,
    name = 'bsp2_test_array'
)

}
\author{
Leonardo Collado-Torres
}
