% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_single.R
\name{job_single}
\alias{job_single}
\title{Build a SGE job script}
\usage{
job_single(name, create_shell = FALSE, queue = "shared",
  memory = "10G", cores = 1L, email = "e", logdir = "logs",
  filesize = "100G", task_num = NULL, tc = 20,
  command = "Rscript -e \\"options(width = 120); sessioninfo::session_info()\\"",
  create_logdir = TRUE)
}
\arguments{
\item{name}{A \code{character(1)} vector with the name of the script. Any spaces
will be replaced by underscores.}

\item{create_shell}{A \code{logical(1)} vector specifying whether to create a
shell file for the script.}

\item{queue}{A \code{character(1)} vector with the name of the SGE queue. Check
how busy a given queue is by running \code{qpic -q queuename}.}

\item{memory}{The amount of memory per core to request in SGE syntax. You
can check how much a current job is utilizing using the \code{qmem} JHPCE command.
For more detail on the memory options, check
\url{https://jhpce.jhu.edu/knowledge-base/how-to/#MemSpec}.}

\item{cores}{The number of cores to request. Note that the total memory
your job will request is \code{cores} multiplied by \code{memory}.}

\item{email}{The email reporting option for the email. For more information
check \url{https://jhpce.jhu.edu/knowledge-base/how-to/#Email}.}

\item{logdir}{The directory for the SGE log files relative to the current
working directory.}

\item{filesize}{The maximum file size in SGE format.}

\item{task_num}{The number of tasks for your job, which will make it into an
array job. If \code{NULL} this is ignored.}

\item{tc}{If \code{task_num} is specified, this option controls the number of
concurrent tasks.}

\item{command}{An example command to start your script.}

\item{create_logdir}{A \code{logical(1)} vector specifying whether to create the
\code{logdir} directory. Note that if \code{logdir} doesn't exist and you submit your
job with \code{qsub}, it will immediately fail.}
}
\value{
A character vector with the script contents. If \code{create_shell} was
specified then it also creates the actual script in the current
working directory.
}
\description{
This function builds a template for a Son of Grid Ending (SGE) job script
including array jobs. Check this blog post by John Muschelli to learn
more about array jobs:
\url{https://hopstat.wordpress.com/2013/11/05/array-and-sequential-cluster-jobs/}.
}
\details{
For a given SGE job that is currently running you can alter
the options using \code{qalter}.
}
\examples{

## A regular job
job_single('jhpce_job', create_logdir = FALSE)

## A regular job with 10 cores on the 'imaginary' queue
job_single('jhpce_job', cores = 10, queue = 'imaginary',
    create_logdir = FALSE)

## An array job
job_single('jhpce_job_array', task_num = 20, create_logdir = FALSE)

}
\author{
Leonardo Collado-Torres
}
