% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_date.R
\name{log_date}
\alias{log_date}
\title{From a SGE log file, find the SGE start and end times}
\usage{
log_date(log_file, tz = "EST", date_lines = c("Job starts",
  "Job ends"))
}
\arguments{
\item{log_file}{The path to a SGE log file.}

\item{tz}{Time zone for your SGE cluster. Defaults to \code{EST} for JHPCE.}

\item{date_lines}{A \code{character(2)} vector with the start and end text lines
used for detecting the start and end times of the job.}
}
\value{
A vector of length 2 with the start and end times
}
\description{
This function searches a log file, such as those created with \code{\link[=job_single]{job_single()}},
for the start and end dates for the SGE job. This information can then be
used to find the \code{qacct} file to retrieve information about the job at hand.
}
\details{
Only the first occurrenses of the start and end time line identifiers
defined by \code{date_lines} will be retrieved.
}
\examples{

## Example log file
bsp2_log <- system.file('extdata', 'logs', 'delete_bsp2.txt', package = 'sgejobs')

## Find start/end dates
log_date(bsp2_log)

## Another example log file
twas_gene_HIPPO <- system.file('extdata', 'logs',
    'compute_weights_full_HIPPO_gene.1.txt', package = 'sgejobs')
log_date(twas_gene_HIPPO)

}
\seealso{
\code{\link[=log_jobid]{log_jobid()}}
}
\author{
Leonardo Collado-Torres
}
