% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_jobid.R
\name{log_jobid}
\alias{log_jobid}
\title{From a SGE log file, find the job id}
\usage{
log_jobid(log_file, jobid_header = "Job id:")
}
\arguments{
\item{log_file}{The path to a SGE log file.}

\item{jobid_header}{A \code{character(1)} vector with the line header that
specifies the SGE ID.}
}
\value{
A \code{character(1)} vector with the job id (first one found in the log
file).
}
\description{
This function searches a log file, such as those created with \code{\link[=job_single]{job_single()}},
for SGE job identifier.
}
\examples{

## Example log file
bsp2_log <- system.file('extdata', 'logs', 'delete_bsp2.txt', package = 'sgejobs')

## Find the job id
log_jobid(bsp2_log)

## Another example log file
twas_gene_HIPPO <- system.file('extdata', 'logs',
    'compute_weights_full_HIPPO_gene.1.txt', package = 'sgejobs')

## Id for the second example
log_jobid(twas_gene_HIPPO)

}
\seealso{
\code{\link[=log_date]{log_date()}}
}
