% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgejobs-package.R
\docType{package}
\name{sgejobs-package}
\alias{sgejobs}
\alias{sgejobs-package}
\title{sgejobs: Helper functions for SGE jobs at JHPCE}
\description{
This package contains helper functions for interacting with the
    JHPCE cluster https://jhpce.jhu.edu/ which is uses Son of Grid Engine (SGE)
    for managing the cluster jobs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LieberInstitute/sgejobs}
  \item Report bugs at \url{https://github.com/LieberInstitute/sgejobs/issues}
}

}
\author{
\strong{Maintainer}: Leonardo Collado-Torres \email{lcolladotor@gmail.com} (0000-0003-2140-308X)

}
\keyword{internal}
